state that LCC is a C89 compiler
The README does not currently state that LCC is a C89 compiler. While this can be inferred based on the book's publication date of 1995, it causes much confusion (see #47). This patch adds a blurb to the README informing users that it is a C89 compiler; a warning to not run C99 should hopefully not be necessary.

Signed-off-by: Amy Parker <apark0006@student.cerritos.edu>